Move Error to routes (#33)
